requirements.txt
streamlit==1.40.2
pandas==2.2.3
numpy==1.26.4
langchain==0.3.19
langchain_core==0.3.49
langgraph==0.3.21
langchain_anthropic==0.3.10
langchain_openai==0.3.7
langchain_huggingface==0.1.2
anthropic==0.49.0
pillow==9.3.0
requests==2.32.3
python-dotenv==1.0.1
#pip install -U craiyon.py
replicate==1.0.4